chore(query): restructure and put behind parameter
